+++ CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `LocationSearch` component now send queries to `vtex.geolocation-graphql-interface` instead of `places-graphql`.
 
 ## [0.10.0] - 2020-09-01
 ### Added
-- `LocationSearch`
+- `LocationSearch`.
 
 ## [0.9.1] - 2020-08-18
 
@@ -37,7 +39,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Prop `hiddenFields` to `PlaceDetails` component.
 
 ### Changed
-- `LocationCountry` now shows the country flag alognside their name.
+- `LocationCountry` now shows the country flag alongside their name.
 
 ### Fixed
 - `LocationCountry` shown with only one country.